+++ runtime/include/chpl-gpu-diags.h
@@ -56,15 +56,15 @@ typedef struct _chpl_gpuDiagnostics {
 #undef _GPU_DIAGS_DECL
 } chpl_gpuDiagnostics;
 
-void chpl_gpu_startVerbose(chpl_bool, chpl_bool, int32_t, int32_t);
-void chpl_gpu_stopVerbose(int32_t, int32_t);
-void chpl_gpu_startVerboseHere(chpl_bool, chpl_bool, int32_t, int32_t);
-void chpl_gpu_stopVerboseHere(int32_t, int32_t);
+void chpl_gpu_startVerbose(chpl_bool, chpl_bool, int64_t, int32_t);
+void chpl_gpu_stopVerbose(int64_t, int32_t);
+void chpl_gpu_startVerboseHere(chpl_bool, chpl_bool, int64_t, int32_t);
+void chpl_gpu_stopVerboseHere(int64_t, int32_t);
 
-void chpl_gpu_startDiagnostics(chpl_bool, int32_t, int32_t);
-void chpl_gpu_stopDiagnostics(int32_t, int32_t);
-void chpl_gpu_startDiagnosticsHere(chpl_bool, int32_t, int32_t);
-void chpl_gpu_stopDiagnosticsHere(int32_t, int32_t);
+void chpl_gpu_startDiagnostics(chpl_bool, int64_t, int32_t);
+void chpl_gpu_stopDiagnostics(int64_t, int32_t);
+void chpl_gpu_startDiagnosticsHere(chpl_bool, int64_t, int32_t);
+void chpl_gpu_stopDiagnosticsHere(int64_t, int32_t);
 void chpl_gpu_resetDiagnosticsHere(void);
 void chpl_gpu_getDiagnosticsHere(chpl_gpuDiagnostics *cd);
